Added load and Assert to objects
A loadfunction with hash assert has been added.  Comparing the old stored hash to a new has made from the json file with the same file name.

In the store function there has been a change where as it now stores a hash in a .txt file instead of storing a pickle file.